+++ utils/linked_list/l_len.c
@@ -6,13 +6,13 @@
 /*   By: aben-ham <aben-ham@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/24 20:47:19 by aben-ham          #+#    #+#             */
-/*   Updated: 2022/01/28 21:26:01 by aben-ham         ###   ########.fr       */
+/*   Updated: 2022/01/28 22:55:02 by aben-ham         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "linked_list.h"
 
-size_t	l_len(t_list *list)
+long	l_len(t_list *list)
 {
 	if (!list || !list->head)
 		return (0);
